Decompose files
